Renamed the service file
